Added features
Added a kill switch.  When triggered it will stop running the CV and gracefully exit.  This is intended to be used as a mechanical lap counter so the car will stop once it passes the finish line.

PIDs were tuned, the motor speed was upped to 60 which is a brisk walk.  The PIDs were then tuned to be a fairly good starting point.  The PIDs for 40 and 50 are still there but commented.

Handles Ctrl + C in the main program.  If you hit Ctrl + C to exit at the wrong time the car could keep going.  This catches the keyboard interrupt and gracefully exits.

Few other minor tweaks.